Strip trailing whitespace with project.core.ui.display_columns
pdm-show uses `project.core.ui.display_columns`, which adds trailing spaces to
all the lines (based on the longest one), which results in them being
displayed/wrapped when the terminal is not wide enough for the longest
line:

```
% pdm show coverage|cat -A
Name:                  coverage                                       …$
Latest version:        6.0.2                                          …$
…
Project URLs:          Documentation: https://coverage.readthedocs.io …$
                       Funding: https://example.com/?utm_source=foobar…$
…
```

This patch just adds `rstrip()`, which gets rid of it, while still
keeping the maybe too long underline for the header (not used with
pdm-show).

Maybe this could use $COLUMNS / the terminal width to limit the length
of the header underline, but that is not really needed / not trivial.