docs: clarify the behavior of includes (#1532)
The behavior of `includes` seems to imply that any files listed will be
*added* to the package itself. However the latter has to be also
explicitly included. This can cause confusion when adding `includes` for
the first time in an already existing package, whereby the package itself
will suddenly disappear from the build. Clarify this in the
documentaion.

Fixes https://github.com/pdm-project/pdm/issues/1531